Fix curve inquirer
