+++ rotkehlchen/chain/ethereum/modules/curve/curve_cache.py
@@ -12,12 +12,11 @@ from rotkehlchen.constants import ONE
 from rotkehlchen.db.drivers.gevent import DBCursor
 from rotkehlchen.errors.misc import NotERC20Conformant
 from rotkehlchen.globaldb.cache import (
-    compute_cache_key,
     globaldb_delete_general_cache,
     globaldb_delete_general_cache_like,
-    globaldb_get_general_cache_keys_and_values_like,
     globaldb_get_general_cache_values,
     globaldb_set_general_cache_values,
+    read_curve_data,
 )
 from rotkehlchen.globaldb.handler import GlobalDBHandler
 from rotkehlchen.logging import RotkehlchenLogsAdapter
@@ -52,31 +51,6 @@ READ_CURVE_DATA_TYPE = tuple[
 ]
 # list of pools that we know contain bad tokens
 IGNORED_CURVE_POOLS = {'0x066B6e1E93FA7dcd3F0Eb7f8baC7D5A747CE0BF9'}
-
-# Using any random address here, since length of all addresses is the same
-BASE_POOL_TOKENS_KEY_LENGTH = len(compute_cache_key([GeneralCacheType.CURVE_POOL_TOKENS, ZERO_ADDRESS]))  # noqa: E501
-
-
-def read_curve_data(
-        cursor: 'DBCursor',
-        pool_address: ChecksumEvmAddress,
-) -> list[ChecksumEvmAddress]:
-    """
-    Reads tokens for a particular curve pool. Tokens are stored with their indices to make sure
-    that the order of coins in pool contract and in our cache is the same. This functions reads
-    and returns tokens in sorted order.
-    """
-    tokens_data = globaldb_get_general_cache_keys_and_values_like(
-        cursor=cursor,
-        key_parts=[GeneralCacheType.CURVE_POOL_TOKENS, pool_address],
-    )
-    found_tokens: list[tuple[int, ChecksumEvmAddress]] = []
-    for key, address in tokens_data:
-        index = int(key[BASE_POOL_TOKENS_KEY_LENGTH:])  # len(key) > BASE_POOL_TOKENS_KEY_LENGTH
-        found_tokens.append((index, string_to_evm_address(address)))
-
-    found_tokens.sort(key=lambda x: x[0])
-    return [address for _, address in found_tokens]
 
 
 def read_curve_pools_and_gauges() -> READ_CURVE_DATA_TYPE:
